Actualización radio enalce.

+++ Shared/RadioEnlace.Shared/Dto/ElevationDto.cs
@@ -5,7 +5,7 @@ namespace RadioEnlace.Shared.Dto
     public class ElevationDto : CoordinateDto
     {
         [JsonProperty("elevation")]
-        public uint Elevation { get; set; }
+        public double Elevation { get; set; }
         [JsonProperty("error")]
         public string Error{ get; set; }
     }
